refactor(PrizeCountMap): matchResults를 EnumMap으로 만들 때 Collectors.groupingBy() 이용
- Collectors.counting()의 return type에 맞춰 PrizeCountMap의 value도 Long으로 변경

+++ src/main/java/lotto/model/prize/Prize.java
@@ -43,7 +43,7 @@ public enum Prize {
 
     }
 
-    public long pickAmount(int count) {
+    public long pickAmount(Long count) {
         return this.amount * count;
     }
